drivers: ambarella: nand: Decouple low-level config from AMBALINK
Configure the hardware to operate as if CONFIG_PLAT_AMBARELLA_AMBALINK
were enabled, even though this may not be the case in our kernel.
This is needed to allow the NAND to be configured in the "standard"
ambalink format, so that we can write to it in the same format as
what stock firmware would eventually expect.

Signed-off-by: evilwombat <evilwombat@server.fake>